requirements.txt
pyubx2==1.2.17
pyserial==3.5
depthai==2.30.0.0
opencv-python==4.8.0.76
pyusb==1.3.1
numpy==1.26.4
pillow==10.2.0
wheel==0.42.0
protobuf==4.25.3
filelock==3.13.1
cffi==1.17.1
setuptools==78.1.0
onnx==1.16.0
packaging==25.0
pytz==2024.1
torch==2.1.0
ultralytics-thop==2.0.14
requests==2.32.3
colorama==0.4.6
PyYAML==6.0.1
charset-normalizer==3.4.2
typing_extensions==4.12.2
scipy==1.12.0
matplotlib==3.8.4
mpmath==1.3.0
sympy==1.12
roboflow==1.1.64
ultralytics==8.3.145
tqdm==4.67.1
torchvision==0.16.0
psutil==5.9.8
networkx==3.3
urllib3==2.2.1
fsspec==2024.6.1
MarkupSafe==2.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
pyrtcm==1.1.6
pynmeagps==1.0.50
certifi==2025.4.26
py-cpuinfo==9.0.0
six==1.17.0
filetype==1.2.0
onnxslim==0.1.53
idna==3.7
python-dotenv==1.0.1
pillow_heif==0.22.0
requests-toolbelt==1.0.0
contourpy==1.2.1
fonttools==4.51.0
pycparser==2.22
pyparsing==3.2.3
Jinja2==3.1.4
cycler==0.12.1
kiwisolver==1.4.5
torchaudio==2.2.2
humanfriendly==10.0
coloredlogs==15.0.1
flatbuffers==25.2.10
timezonefinder==6.5.9
h3==4.2.2